+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2024-07-01
 - Interactive items that can be picked up, put down, or even stolen from other mice
+- Items can move in between rooms
 - Minimap now shows other players and items
 - "enable sounds" option in settings panel
 - Chat keystrokes are now audible & visible